Fixing issues with text overflow
