+++ ui/app/src/components/SiteConfigurationManagement/styles.ts
@@ -15,7 +15,6 @@
  */
 
 import { createStyles, makeStyles } from '@material-ui/core/styles';
-import palette from '../../styles/palette';
 
 export const useStyles = makeStyles((theme) =>
   createStyles({
@@ -90,30 +89,6 @@ export const useStyles = makeStyles((theme) =>
     },
     itemSkeletonText: {
       height: '20px'
-    }
-  })
-);
-
-export const useResizeableStyles = makeStyles((theme) =>
-  createStyles({
-    resizeHandle: {
-      width: '2px',
-      minWidth: '2px',
-      margin: '0px 5px',
-      cursor: 'ew-resize',
-      padding: '4px 0 0',
-      backgroundColor: 'rgba(0, 0, 0, 0.12)',
-      transition: 'width 200ms',
-      '&:hover': {
-        width: '4px',
-        visibility: 'visible',
-        backgroundColor: palette.blue.tint
-      }
-    },
-    resizeHandleActive: {
-      width: '4px',
-      visibility: 'visible',
-      backgroundColor: palette.blue.tint
     }
   })
 );
